backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
numpy
Pillow
groq
python-dotenv
